Move internal Context to bona fide libwallet type
