+++ wallet/src/libwallet/internal/mod.rs
@@ -24,6 +24,5 @@
 pub mod keys;
 pub mod restore;
 pub mod selection;
-pub mod sigcontext;
 pub mod tx;
 pub mod updater;
